requirements.txt
fastapi
uvicorn[standard]
httpx
python-telegram-bot==20.6
wikipedia-api